requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
openai>=1.30.0
python-dotenv>=1.0.1
